requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.0.5=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py312ha883a20_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.3.11=h06a4308_0
decorator=5.1.1=pyhd3eb1b0_0
django-widget-tweaks=1.5.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h6a678d5_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipython=8.20.0=py312h06a4308_0
jedi=0.18.1=py312h06a4308_1
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
matplotlib-inline=0.1.6=py312h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
ncurses=6.4=h6a678d5_0
numexpr=2.8.7=py312hf827012_0
numpy=1.26.3=py312hc5e2394_0
numpy-base=1.26.3=py312h0da6c21_0
openssl=3.0.13=h7f8727e_0
pandas=2.1.4=py312h526ad5a_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pip=23.3.1=py312h06a4308_0
plotly=5.9.0=py312h06a4308_0
prompt-toolkit=3.0.43=py312h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=py312h06a4308_1
python=3.12.1=h996f2a0_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py312h06a4308_0
readline=8.2=h5eee18b_0
setuptools=68.2.2=py312h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.2=py312h06a4308_1
tk=8.6.12=h1ccaba5_0
traitlets=5.7.1=py312h06a4308_0
tzdata=2023d=h04d1e81_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.41.2=py312h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
